requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==5.0.2
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
facebook-sdk==3.1.0
frozenlist==1.4.1
google-auth==2.27.0
idna==3.6
jwcrypto==1.5.4
multidict==6.0.5
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
twilio==8.13.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
yarl==1.9.4
